Replace public class AbstractLabelWiseLoss with private class LabelWiseLoss.
